requirements.txt
anyio==4.10.0
asgiref==3.9.1
autopep8==2.3.2
blinker==1.9.0
certifi==2025.8.3
click==8.2.1
et_xmlfile==2.0.0
Flask==3.0.3
gunicorn==22.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
pycodestyle==2.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.0.6
